Adds a ArgumentParser object to process flags and cleans up code more.
I recently discovered `ArgumentParser.parse_known_args`, which fixes the
problems that stopped me from using the argparse module for this in the first
place. I'm going to try using argparse for this project for now since I've
found this.

Error introduced last commit where a variable was renamed in one place and
not another has been fixed.